Keep tracking rate central difference in double
Mount::poll() derives the Axis1 and Axis2 tracking rates as a central
difference over DiffRange2, but copied the double Coordinate members into
float locals before the subtraction. Differencing two values of order
1 radian to extract 5.8e-4 radians, after rounding both to single
precision, discards most of the significant digits of the result.

The Coordinate members are double throughout, so the precision is
available right up to the point it was thrown away. rate1 and rate2 stay
float: the loss was in the subtraction, not the result.

No effect on HAL_NO_DOUBLE_PRECISION targets, where double is float and
DiffRange is already widened to 30 arc-minutes to compensate.

+++ src/telescope/mount/Mount.cpp
@@ -525,7 +525,8 @@ void Mount::poll() {
   }
 
   // transfer to variables named appropriately for mount coordinates
-  float aheadAxis1, aheadAxis2, behindAxis1, behindAxis2;
+  // these stay double, the differences below are ~5.8e-4 rad from ~1 rad coordinates
+  double aheadAxis1, aheadAxis2, behindAxis1, behindAxis2;
   if (transform.mountType == ALTAZM) {
     transform.equToHor(&ahead);
     aheadAxis1 = ahead.z;
